Move rubriken - aktivzitierung to separate file
